+++ myapp/index.js
@@ -1,0 +1,6 @@
+function countrySelect(country) {
+    var selector = document.getElementById("country-" + country);
+    if (selector != null) {
+        selector.selected = true;
+    }
+}